pyproject.toml
[project]
name = "boondoggle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "pillow>=12.0.0",
    "xlsxwriter>=3.2.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.4",
]

[tool.ruff.lint]
# Honestly going a bit overboard here as I've just worked out where these come from
# https://docs.astral.sh/ruff/rules/
extend-select = [ 
    "FAST",
    "ASYNC",
    "I",
    "B",
    "A",
    "C4",
    "EM",
    "PT",
    "SIM",
    "PERF",
]
